Add NoteDialogButton widget
Introduce a new widget `NoteDialogButton` for handling the save functionality in the note dialog.

This widget uses a `BlocBuilder` to listen to the `AddOrUpdateNoteCubit` state and displays a `CustomElevatedButton` that triggers the save action when pressed.

The button's loading state is managed based on the cubit's state.

The save action validates the form and either updates an existing note or adds a new one based on the presence of a note model.